external integration test ready for time-based. moving to work on unit test level
